#2127 - refactored MarginPositions component, #1098 - added sorting collaterals by debt
