+++ app/components/Account/CollateralTable.jsx
@@ -1,0 +1,255 @@
+import React from "react";
+import MarginPosition from "./MarginPosition";
+import ChainTypes from "../Utility/ChainTypes";
+import BindToChainState from "../Utility/BindToChainState";
+import AssetWrapper from "../Utility/AssetWrapper";
+import {ChainStore} from "bitsharesjs";
+import Translate from "react-translate-component";
+import counterpart from "counterpart";
+import {List} from "immutable";
+import TranslateWithLinks from "../Utility/TranslateWithLinks";
+import Immutable from "immutable";
+
+const alignRight = {textAlign: "right"};
+const alignLeft = {textAlign: "left"};
+
+/**
+ * Renders whole margin list including call orders and placeholders
+ */
+class ListGenerator extends React.Component {
+    static propTypes = {
+        callOrders: ChainTypes.ChainObjectsList
+    };
+
+    constructor(props) {
+        super(props);
+
+        this.state = {
+            assetsPropsCount: 0,
+            callOrdersCount: 0
+        };
+    }
+
+    static getDerivedStateFromProps(nextProps, prevState) {
+        // if we got new bitasset info
+        // or call orders changed
+        // then recalculate "cache" of maring items
+
+        const callOrdersJson = JSON.stringify(nextProps.callOrders);
+
+        if (
+            prevState == null ||
+            prevState.assetsPropsCount != nextProps.bitAssets.length ||
+            prevState.ordersJson != callOrdersJson
+        ) {
+            let assets = new Array();
+            // construct map of bitassets
+            let assetsMap = {};
+            // index to track asset info
+            let index = 0;
+
+            nextProps.bitAssets.forEach(o => {
+                assetsMap[o.get("id")] = index++;
+                assets.push({
+                    asset: o,
+                    has_margin_order: false
+                });
+            });
+
+            if (nextProps.callOrders.length > 0) {
+                // add to iterated bitasets items that wasn't listed by default (component's callOrders prop)
+                nextProps.callOrders.forEach(o => {
+                    let assetId = o.getIn(["call_price", "quote", "asset_id"]);
+
+                    if (assetsMap[assetId] == null) {
+                        let newAssetInfo = ChainStore.getObject(assetId);
+                        if (typeof newAssetInfo != "undefined") {
+                            assetsMap[assetId] = index++;
+                            assets.push({
+                                asset: newAssetInfo,
+                                has_margin_order: true,
+                                order: o
+                            });
+                        }
+                    } else {
+                        // mark as margin order
+                        assets[assetsMap[assetId]].has_margin_order = true;
+                        assets[assetsMap[assetId]].order = o;
+                    }
+                });
+            }
+
+            return {
+                assetsPropsCount: nextProps.bitAssets.length,
+                assets: assets,
+                ordersJson: callOrdersJson
+            };
+        }
+
+        return prevState;
+    }
+
+    render() {
+        let {account, callOrders} = this.props;
+        callOrders = callOrders.filter(o => !!o);
+
+        let rows = this.state.assets
+            .sort((a, b) => {
+                // a,b - order ChainObject's
+                // could be done via component's property `order_by`
+
+                let aVal = a.has_margin_order ? a.order.get("debt") : 0;
+                let bVal = b.has_margin_order ? b.order.get("debt") : 0;
+
+                return bVal - aVal;
+            })
+            .map(a => {
+                let debtAsset,
+                    collateralAsset,
+                    order_id = null;
+
+                // user has margin on this asset
+                if (a.has_margin_order) {
+                    let order = a.order;
+
+                    debtAsset = order.getIn([
+                        "call_price",
+                        "quote",
+                        "asset_id"
+                    ]);
+                    collateralAsset = order.getIn([
+                        "call_price",
+                        "base",
+                        "asset_id"
+                    ]);
+
+                    order_id = order.get("id");
+                } else {
+                    debtAsset = a.asset.get("id");
+                    collateralAsset = a.asset.getIn([
+                        "bitasset",
+                        "options",
+                        "short_backing_asset"
+                    ]);
+                }
+
+                return (
+                    <MarginPosition
+                        key={a.asset.get("id")}
+                        object={order_id}
+                        account={account}
+                        debtAsset={debtAsset}
+                        collateralAsset={collateralAsset}
+                        {...this.props}
+                    />
+                );
+            });
+
+        return <tbody>{rows}</tbody>;
+    }
+}
+
+ListGenerator = BindToChainState(ListGenerator);
+ListGenerator = AssetWrapper(ListGenerator, {
+    propNames: ["bitAssets"],
+    defaultProps: {
+        bitAssets: [
+            "1.3.103",
+            "1.3.113",
+            "1.3.120",
+            "1.3.121",
+            "1.3.958",
+            "1.3.1325",
+            "1.3.1362",
+            "1.3.105",
+            "1.3.106"
+        ]
+    },
+    asList: true
+});
+
+const CollateralTable = ({
+    callOrders,
+    account,
+    className,
+    children,
+    preferredUnit
+}) => {
+    return (
+        <table className={"table table-hover " + className}>
+            <thead>
+                <tr>
+                    <th style={alignLeft}>
+                        <Translate content="explorer.asset.title" />
+                    </th>
+                    <th style={alignRight}>
+                        <Translate content="exchange.balance" />
+                    </th>
+                    <th style={alignRight}>
+                        <Translate content="transaction.borrow_amount" />
+                    </th>
+                    <th style={alignRight} className="column-hide-medium">
+                        <Translate content="transaction.collateral" />
+                    </th>
+                    <th>
+                        <div
+                            className="tooltip inline-block"
+                            data-place="top"
+                            data-tip={counterpart.translate(
+                                "tooltip.coll_ratio"
+                            )}
+                        >
+                            <Translate content="borrow.coll_ratio" />
+                        </div>
+                    </th>
+                    <th>
+                        <TranslateWithLinks
+                            noLink
+                            string="account.total"
+                            keys={[
+                                {
+                                    type: "asset",
+                                    value: preferredUnit,
+                                    arg: "asset"
+                                }
+                            ]}
+                        />
+                    </th>
+                    <th style={alignRight} className="column-hide-small">
+                        <div
+                            className="tooltip inline-block"
+                            data-place="top"
+                            data-tip={counterpart.translate(
+                                "tooltip.call_price"
+                            )}
+                        >
+                            <Translate content="exchange.call" />
+                        </div>
+                    </th>
+                    <th style={alignRight} className="column-hide-small">
+                        <Translate content="exchange.price" />
+                    </th>
+                    <th className="column-hide-small" style={alignLeft}>
+                        <Translate content="explorer.assets.units" />
+                    </th>
+                    <th style={{textAlign: "center"}}>
+                        <Translate content="exchange.market" />
+                    </th>
+                    <th>
+                        <Translate content="borrow.adjust_short" />
+                    </th>
+                    <th>
+                        <Translate content="transfer.close" />
+                    </th>
+                </tr>
+            </thead>
+            <ListGenerator
+                account={account}
+                callOrders={Immutable.List(callOrders)}
+            />
+            <tbody>{children}</tbody>
+        </table>
+    );
+};
+
+export default CollateralTable;
